+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.3.0
+- add: creation time display
+- add: sort by creation time
+- add: randomly select
+- fix: fix some bugs
+- add: download list window
+- add: an error dialog box will pop up if the user opens the directory "left4dead2/addons"
 # v0.2.0
 - improve: AddonNodeView
 - fix: WorkshopVpkAddon.FullVpkFilePath is sometimes not updated